feat: pre-deploy audit fixes, catalog, audit log and UX improvements
## Segurança e correções críticas (P0–P1)
- Registra rota de refresh de token JWT (TokenRefreshView)
- Adiciona campo consumer_id obrigatório no ConsumptionForm
- Remove ConsumerIdAuthentication do escopo global
- Exige autenticação em todos os endpoints de lotes
- Corrige colunas trocadas (Unidade/Estoque mínimo) na tabela de Produtos
- DEBUG e ALLOWED_HOSTS passam a ler variáveis de ambiente

## Melhorias de arquitetura (P1–P2)
- FloatField → DecimalField(10,3) em Batch, Consumption e Product (migrations)
- on_delete=CASCADE → PROTECT em Batch.product e Consumption.batch (migrations)
- Paginação padrão configurada (PageNumberPagination, PAGE_SIZE=50)
- JWT access token reduzido de 7 dias para 15 minutos
- django-cors-headers instalado e configurado
- runserver substituído por gunicorn com 2 workers
- Testes de consumo atualizados para JWT + consumer_id no payload

## Melhorias de UX (P3)
- Botões Editar e Excluir na tabela de Produtos (ADMIN/MANAGER)
- ProductForm suporta modo edição via PATCH (/produtos/:id/editar)
- PrivateRoute verifica expiração real do JWT
- RoleRoute criado para proteção por role; /usuarios restrito a ADMIN
- Toast de sucesso em ProductForm e BatchForm antes do redirect
- Formulário de lote bloqueia datas passadas (min date + validação backend)
- Renomeia serizalizers.py → serializers.py em products
- Tratamento de ProtectedError com HTTP 409 em produtos e lotes

## Novas features
- Página /catalogo (ADMIN + MANAGER): CRUD inline de Categorias e Unidades
- Exclusão de lotes habilitada para ADMIN/MANAGER com ProtectedError tratado
- Exclusão de consumos habilitada para ADMIN/MANAGER com restauração atômica do estoque
- Filtro de linhas por página (10/25/50/100) em todas as tabelas
- Auditoria completa com django-auditlog: rastreia criações, edições e exclusões
  de Usuários, Produtos, Categorias, Unidades, Lotes e Consumos
- Página /auditoria (ADMIN): filtros por ação, modelo, usuário e período

+++ backend/products/models.py
@@ -20,7 +20,7 @@ class Product(BaseModel):
     name = models.CharField(max_length=150)
     category = models.ForeignKey(Category, on_delete=models.PROTECT)
     unit = models.ForeignKey(Unit, on_delete=models.PROTECT)
-    minimum_stock = models.FloatField(default=0)
+    minimum_stock = models.DecimalField(max_digits=10, decimal_places=3, default=0)
 
     def __str__(self):
         return self.name
